requirements.txt
# ─── PDF Parsing ──────────────────────────────────────────────────────────────
PyMuPDF>=1.24.0
pytesseract>=0.3.10   # OCR for image-embedded text; requires system Tesseract binary
Pillow>=10.0.0        # Image handling for OCR pipeline

# ─── Chunking & LLM orchestration ────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0

# ─── Vector Store ─────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Embeddings & Reranking ───────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── Sparse Retrieval ─────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ─── UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ─── OpenAI SDK ───────────────────────────────────────────────────────────────
openai>=1.30.0

# ─── Config ───────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Numerics ─────────────────────────────────────────────────────────────────
numpy>=1.26.0

# ─── Evaluation (Phase 5) ─────────────────────────────────────────────────────
scikit-learn>=1.4.0

# ─── Progress bars ────────────────────────────────────────────────────────────
tqdm>=4.66.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
